python binding require swig
Require SWIG. This solves the issue of the optional
dependencies. Namely, when we build the generated files
on another system without SWIG the dependencies have to
match the system where the bindings were generated which
defeats the purpose of optional dependencies. Python
binding are still optional, but to compile them you need
SWIG.