flat python module
in the beginning I wanted to use the same python module
layout that we do in our C++ libraries. specifically
teca
 +- core
 +- data
 +- alg
 +- io
where io and alg each depend on both core and data. This
layout doesn't work well with SWIG, which embeds complete
copies of core and data in both alg and io modules. This
duplicate embedding bloated the python modules quite a bit
and it was not clear the NumPy double free issues we are
seeing are caused by NumPy symbol table conflicts. The flat
module eliminates that potential, and the generated code is
significantly smaller, as are the binaries and the project
builds faster as well.