teca_event_filter to teca_table_region_mask
isolate the geographic filtering capability of the event_filter
into an algorithm that generates a mask. the mask can then be
used in conjunction with the table_remove_rows algorithm to
remove rows based on the geographic locations. Adds a regression
test for the new algorithm. other functionality in the event_filter
is replaced by table_remove_rows.