add teca_parser utility
The parser utility is comprised of a infix parser that converts
expression to postfix and a postfix evaluator. Both make use of
the following abstractions: abstract tokenizer type for identifying
operators, an abstract operator type for applying operators, and
an abstract operand type for servging up operands (variables and
constants).

The test implementation supports the most common logical and numeric
operators (unary,binary,ternary) of C++ and typed constants.