virtualize teca_algorithm ports/connections
make the set input connection and get output port
virtual so that derived classes can override the default
behavior. the driving use case for this change is
the storm size algorithm which needs to run a user
specified pipeline in it's metatdata reporting phase.
the class spoofs a 2 input algorithm, the first input
is ssaved in a member variable and executed in the
metadata phase, the second input is connected as usual.