fix teca_binary_stream size reporting issue
a recurring problem has been that the stream position was
used to record it's size. both read and write used the pointer.
thus if you rewind the stream to read it you lost it's size.

this patch adds read and write positions to the stream's
internal state as well as API to set the current positions. The
stream's size is still computed from the write position.