teca_cf_reader cache metadata
this patch adds disk based metadata caching to the CF-2
reader. the cache is associated with the list of nc files
and path where they are located. it is written to either
the path where the nc files are, the user's $HOME dir from
the environment variable, or the current working dir.

this seeks to address the performance issues we found with
scanning NetCDF's unlimitted dimensions on large datasets.