teca_thread_pool bind threads to unique cores
use the cpuid instruction to detect cpu topology, physical cores
and hyperthreads. When possible we spread thread pool over unique
cores on a node taking into account other MPI ranks running on the
node. When the user requests more than the number of physical cores
available on a node, we bind to unique hyper threads. When the user
requests more than the number of hyperthreads on the node, we bind
to least used hyperthread. only implemented for intel.