improve Python exceptions coming from the C-API
In Python in order for a C/C++ function to throw an exception 2 things
must happen: 1) set the error variable, 2) return NULL. This patch adds
retrun values to extension functions that exist only in the SWIG/glue
layer so that step 2 is possible. These functions now return Py_None for
success and nullptr for error. This fixes a long standing bug that an
exceptions in these functions do not get reported immediately.

A second issue addressed is that Not all of our extensions can return
PyObject's. Many return std::shared_ptr's SWIG's handling of these is
that a std::shared_ptr pointing to a nullptr evaluates to Py_None. Thus
2 above is impossible to acheive. In these cases we need to set the
error variable and then directly invoke the interpreter before
returning. The new TECA_PY_ERROR macro handles this and sends the
standard TECA_ERROR message through one of the built in Python
exceptions.