add teca_normalize_coordinates
The new algorithm makes sure coordinates from various sources are
presented the same way in TECA. The normalize filter currently detects
descending coordinate axes and makes the ascending. When coordinates are
ascending the algorithm is a no-op and a shollow copy of the input data
is forwarded downstream. The patch installs the algorithm in all
existing apps and tests and include the following tests.

tests: the CF reader test excercises the algorithm on NOAA CFSR
re-analysis data. this doesn't do a diff though. A set of test verifies
permutations of axes flips and bounds based request subsetting. the
2d_component_area test is used, because flipped coordinate axes will
result in negative areas