add teca_normalize_coordinates
The new algorithm makes sure coordinates from various sources are
presented the same way in TECA. The normalize filter currently
detects descending coordinate axes and makes the ascending. When
coordinates are ascending the algorithm is a no-op and a shollow
copy of the input data is forwaqrded downstream. The patch installs
the algorithm in all existing apps and tests and includes a new test
that excercises the algorithm on NOAA CFSR re-analysis data.