cartesian mesh source
add an algorithm that generates and serves up a Cartesian mesh. This
comes up a lot in testing. Previously we have been using the dataset
source or programmable algorithm. The new class handles repetative tasks
such as pipeline executive, metadata reporting, and subsetting. It also
handles generation of user defined scalar fields of the form f(x,y,z,t).
This makes it much easier to generate meshes on the fly.