aquire GIL in all Python extensions
This change makes sure the GIL is aquired in certain extension
code that we did not envision to be used in multi threaded execution.
With pipelines becoming more complex, this code may end up running
in multiple threads and thus we need to aquire the GIL before calling
Python C-API. Moving forward we should always aquire the GIL to be
on the safe side.