cf_reader file series without time dimension
make a dummy time axis, this enables parallelization over file sets
that do not have time dimension. However, there is no guarantee on the
order of the dummy axis to the lexical ordering of the files and there
will be no calendaring information. As a result many time aware algorithms
will not work.