build requirements flags
add some flags that can be passed on the cmake command line to
specify which dependencies are to be required by this build.
typically we configure features based on the set of dependencies
founhd, however with the PyPi package we want things to error out
when core dependencies - such as udunits are not found.