optionally disable MPI collectives in thread pool construction
this enables us to nest the thread pools as MPI collectives cannot be
used inside a thread. MPI collectives are used to determine the optimal
thread pool size and coordinate binding of threads to cores.  This patch
defers construction, until a derived class or application explicitly
sets a thread pool size, giving one the chance to disable MPI if needed.