make the MPI communciator setable (issues #202 #183)
Adds API to teca_algorithm for setting/getting a communicator. The
default is MPI_COMM_WORLD. A communicator passed into the set API is
duplicated.

In addition to enabling pipelines to run in an isolated communication
space, this solves some issues when running nested map-reduce.

A number of algortihms make collective queries the MPI run time for info
about the number and/or locations of MPI ranks, and then make
configuration and/or load balancing decisions based on the info. These
collectives can cause deadlocks when run from inside threads in a nested
pipeline. In those cases one can set the communicator for each stage of
the nested pipeline to MPI_COMM_SELF making all the MPI collectives
local operations.