always call reduce override
Prior to this when there is only a single dataset reduce override is not
called, and the single dataset is returned. However, this strategy breaks
for reductions that transform the data. If reduce is not called the data
would not be transformed, leading to different output when a rank only
has a single dataset.