support negative testing
Tests relying on detecting an error to pass can now be enabled
by passing `WILL_FAIL` flag to teca_add_test. This will cause
ctest to pass when `ERROR` is found in the test output. See
test_dataset_diff, for an example which can be run in positive
(verifies 2 datasets are the same) or negative(verifies 2 datasets
are different)  mode.