MPI error handler
Add mechanism to install an MPI error handler callback. The handler must
be attached to a communicator and prints a stack trace. It is currently
used in regression tests but not the applications. It would be fine to
add it to the applications as well. Add a regression test that
excercises the handler.