algorithm_executive requests index 0
prior to this the default request was an effictively empty request.
this changes to always request index 0 using the index_request_key.
this is an easy way to handle the common use case where the same dataset
needs to be served up across MPI ranks and makes it so the default
executive is compatible with all of our existing data sources.