isloate pipeline below the reduction
a reduction changes the structure of the dataset being passed
from a massive dataset with many pieces to a small one with
a single piece. pipeline stages running below the reduction
need different control metadata. This control metadata cannot
be passed up stream. hence the need for isolating the pipelines
above and below teh reduction.