add cf2_writer
adds a NetCDF CF2 writer for Cartesian meshes. The writer can write any
number of time steps per file. It will drive the upstream in order to
gather the time steps it needs for each file. The writer is intended to
be used in parallel with MPI and/or threads. Use tthe index_executive to
partitions files across ranks. Adds tests that exercise the writer in
various configurations.