add thread parallel streaming mode
this lets results from the teca_threaded_algorithm be processed as they
become ready rather than all at once when all are ready.  results are
processed by the thread driving the pipeline. A new streaming flag
passed to ::execute and virtual method indicates if it is safe to use
MPI or other non-threadsafe operations. The flag is set to 0 when it is
safe to use MPI etc. Includes a regression test.