make the python_algorithm polymorphic
the author of a new algorithm would now have to override report,
request, and execute methods instead of get'ers that return functions
implementing these.

the get_X_callback methods remain in the python_algorithm and use the
closeruse trick but now return a function that calls the above class
methods (report, request, execute).

from a user's pov this should be a simplification as they no longer need
to know about the closure trick.

this change makes it possible to use polymorphism to provide default
implementations that could be called after some preprocesing of the
inputs. This could be useful in teca_model_segmentation and its child
classes.