add a finalize stage to reductions
finalize stage is needed for some types of reduction
such as the average operation employed by the Bayesian AR
detector. Implemented in teca_index_reduce as a new
virtual method that is invoked after the remote data has
been reduced. The default passes the input through.
Added the machinery to make this work in the programmable
reduce as well, and exposed it in Python.