add teca_priority_queue
Adds indirect priority queue that supports random access
modification of priorities in-situ. users of the queue
intialize the queue with keys and an accessor callable
that converts keys to priority values. support for generic
key type is provided as well as an optimized specical case
for integer ordinal keys. Keys are arranged according to
the specified predicate (std::less is the default). pop/peak
operations return the key with the "highest" priority.