support NetCDF parallel I/O in the CF writer
enables to writer to load balance across time steps and then put local
steps in a collection of files the number of which are determined at
run time by the number of time steps to process and the steps per file
parameter. many ranks may be writing to the same file. data may be
written in arbitrary order, it will land in the correct position in the
correct file.

This is a major restructuring. Design constraints include that certain
NetCDF operations are collective which forces them to occur during the
phases of pipeline where such collective operations are safe. For this
reason one must know the mesh size, list of arrays, each arrays type and
size if it is not mesh oriented during report phase of execution. Also
certain NetCDF features that we used in the past, specifically unlimited
dimension and compression, are incompatible with our execution patterns
because they require collective writes and it is not safe to do
collective writes in the streaming execution pattern.