add teca_threaded_programmable_algorithm
This class is similar to the teca_programmable_algorithm but rather than
inheriting from teca_algorithm is inherits from teca_threaded_algorithm
giving it thread parallel execution. This class is most useful for
scenarios where multiple requests per execute are made. The requests are
satsified by executing the upstream pipeline in parallel using a thread
pool. Data can be processed as it is generated using the streaming mode.
Because the thread pool has MPI collectives care must be taken when
using to avpoid dead locks. This is why this class exists instead of
just extending the programmable_algorithm.