deal with missing values in temporal_reduction
The set_use_fill_value  method controls how invalid or missing values are
teated.  When set to 1, NetCDF CF fill values are detected and handled.
This is the default. If it is known that the dataset has no invalid or
missing values one may set this to 0 for faster processing. By default the
fill value will be obtained from metadata stored in the NetCDF CF file
(_FillValue). One may override this by explicitly calling set_fill_value
method with the desired fill value.

For minimum and maximum operations, at given grid point only valid values
over the interval are used in the calculation.  if there are no valid
values over the interval at the grid point it is set to the fill_value.

For the averaging operation, during summation missing values are treated
as 0.0 and a per-grid point count of valid values over the interval is
maintained and used in the average. Grid points with no valid values over
the inteval are set to the fill value.