add valid_value_mask
the valid_value_mask algorithm generates a mask by scanning for the
value of the NetCDF _FillValue attribute in specified data arrays.
The list of arrays to process may be provided explicitly or requested
from down stream by appending "_valid" to the name of an existing array.
includes a test and Python bindings.