array_attributes support for _FillValue
adds an optional atrtribute for fill value. one must also set the
associated flag to enable it. internally c++17 variant is used
and the type of the contained value must be the same as the array
it will be used with, otherwise a NetCDF error will occur when
writing the data. Includes Python bindings. In Python no flag
is used, pass None to indicate no fill value.