add spatial executive
this executive partitions a time series of 3D Cartesian mesh based data
into a 2D space where axes consit of the spatial partitioning and the
temporal partitioning. The number of elements in the space and time
dimensions of the partitioning be runtime specified. The resulting
partitions are assigned to MPI ranks cyclically such that MPI ranks
processes all time temporal partitions associated with assigned spatial
partition.  includes a test and Python bindings.