fix thread parameters more ranks than cores
in the case that there are more MPI ranks than physical cores,
warn about performance and return an error. 1 thread is reported
and the core id that the code is running on is returned  in the
affinity map. it is not recommended to bind threads to cores in
this case.

we've only encountered this issue on Travis-CI in docker containers.