temporal_reduction fix off by one
The time_step_of index search was set to return the step less than or
equal to the given time point on the low side of the interval. This was
resulting in the last step of the month(day/season) being duplicated.
This patch sets the search to be great or equal on the low side of the
interval.