integrated_vapor_transport layer thickness
support computing layer thickness between vertical level interfaces.
Note: the original approach computed thickness between level centers and
itegrated using trapezoid method while the new approach uses a first
order method. Since level interfaces are not specified with the dataset,
the user must provide correct values for the surface and top level. The
defaults are 101325 Pa and 0 Pa.  Note: when this method is used, the
answer will be wrong if the surface and top layer values do not match
the data in the file or if the vertical axis is cropped. There is no
error checking you will get a wrong answer. For this reason the new
approach is disabled by default.